+++ src/components/common/ComponentBox/index.ts
@@ -1,0 +1,3 @@
+import ComponentBox from './ComponentBox';
+
+export default ComponentBox;
