+++ src/components/page/Home/Home.page.tsx
@@ -1,0 +1,13 @@
+import { VFC } from 'react';
+import Layout from '@/components/common/Layout';
+import Home from './Home';
+
+const HomePage: VFC = () => {
+  return (
+    <Layout>
+      <Home />
+    </Layout>
+  );
+};
+
+export default HomePage;
